OneClickSummary/requirements.txt
python-dotenv
uvicorn
fastapi
firebase-admin
groq
pydantic
